% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{create_endpoint}
\alias{create_endpoint}
\title{create_endpoint}
\usage{
create_endpoint(
  path,
  description = NULL,
  workspace = NULL,
  public_within_org = NULL,
  visibility = NULL
)
}
\arguments{
\item{path}{Path for the endpoint.}

\item{description}{Description of the endpoint.}

\item{workspace}{Workspace under which the endpoint with name name exists. If not provided, the current
user's personal workspace will be used.}

\item{public_within_org}{If creating an endpoint in an organization's workspace: True
for public, False for private. In older backends, default is
private; in newer backends, uses the org's settings by default.}

\item{visibility}{Visibility to set when creating this endpoint. If not provided, an
appropriate default will be used. This parameter should be
preferred over public_within_org.}
}
\value{
:class:`verta.registry._entities.model.RegisteredModel`
}
\description{
Attaches an endpoint to this Client.
}
\details{
An accessible endpoint with name `name` will be created and initialized with specified metadata parameters.
}
