% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{create_experiment}
\alias{create_experiment}
\title{create_experiment}
\usage{
create_experiment(name = NULL, desc = NULL, tags = NULL, attrs = NULL)
}
\arguments{
\item{name}{Name of the Experiment. If no name is provided, one will be generated.}

\item{desc}{Description of the Experiment.}

\item{tags}{Tags of the Experiment.}

\item{attrs}{Attributes of the Experiment.}
}
\value{
:class:`~verta._tracking.experiment.Experiment`
}
\description{
Creates a new Experiment under the currently active Project.
}
\details{
Experiment with name `name` will be created and initialized with specified metadata parameters.
}
