% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{create_project}
\alias{create_project}
\title{create_project}
\usage{
create_project(
  name = NULL,
  desc = NULL,
  tags = NULL,
  attrs = NULL,
  workspace = NULL,
  public_within_org = NULL,
  visibility = NULL
)
}
\arguments{
\item{name}{Name of the Project. If no name is provided, one will be generated.}

\item{desc}{Description of the Project.}

\item{tags}{Tags of the Project.}

\item{attrs}{Attributes of the Project.}

\item{workspace}{Workspace under which the Project with name name exists. If not provided, the current}

\item{public_within_org}{If creating a Project in an organization's workspace: True for}

\item{visibility}{Visibility to set when creating this project. If not provided, an}
}
\value{
:class:`verta._tracking.project.Project`
}
\description{
Creates a new Project.
}
\details{
A Project with name `name` will be created and initialized with specified metadata parameters. If an Experiment is already attached to this Client, it will be detached.
}
