% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{create_registered_model}
\alias{create_registered_model}
\title{create_registered_model}
\usage{
create_registered_model(
  name = NULL,
  desc = NULL,
  labels = NULL,
  workspace = NULL,
  public_within_org = NULL,
  visibility = NULL
)
}
\arguments{
\item{name}{name}

\item{desc}{desc}

\item{labels}{labels}

\item{workspace}{workspace}

\item{public_within_org}{public_within_org}

\item{visibility}{visibility}
}
\value{
:class:`~verta.registry._entities.model.RegisteredModel`
}
\description{
Creates a new Registered Model.
}
\details{
A registered_model with name `name` does will be created and initialized with specified metadata parameters.
}
