% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_io.R
\name{createDockerfile}
\alias{createDockerfile}
\title{Create Dockerfile From Saved Model Object}
\usage{
createDockerfile(
  model_file_location,
  required_packages = c("forecast"),
  plumber_template_location = file.path(path.package("vertaReticulateClient"),
    "plumber_sample.R", additional_objects_file_location = "additional_objects.RData"),
  runner_template_location = file.path(path.package("vertaReticulateClient"),
    "to_run.R")
)
}
\arguments{
\item{model_file_location}{location of the model file}

\item{required_packages}{required packages to hold}

\item{plumber_template_location}{location to the plumber api script, by default the
one from the package is used. Leave it as-is.}

\item{runner_template_location}{location of the runner template script, by default
the one from the package distribution is used. Leave as is, normally.}
}
\value{

}
\description{
Create Dockerfile From Saved Model Object
}
