% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{download_endpoint_manifest}
\alias{download_endpoint_manifest}
\title{download_endpoint_manifest}
\usage{
download_endpoint_manifest(
  download_to_path,
  path,
  name,
  strategy = NULL,
  resources = NULL,
  autoscaling = NULL,
  env_vars = NULL,
  workspace = NULL
)
}
\arguments{
\item{download_to_path}{Local path to download manifest YAML to.}

\item{path}{Path of the endpoint.}

\item{name}{Name of the endpoint.}

\item{strategy}{Strategy (direct or canary) for updating the endpoint.}

\item{resources}{Resources allowed for the updated endpoint.}

\item{autoscaling}{Autoscaling condition for the updated endpoint.}

\item{env_vars}{Environment variables.}

\item{workspace}{Workspace for the endpoint. If not provided, the current user's}
}
\value{
downloaded_to_path
}
\description{
Downloads this endpoint's Kubernetes manifest YAML.
}
\details{

}
