% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{find_datasets}
\alias{find_datasets}
\title{find_datasets}
\usage{
find_datasets(
  dataset_ids = NULL,
  name = NULL,
  tags = NULL,
  sort_key = NULL,
  ascending = FALSE,
  workspace = NULL
)
}
\arguments{
\item{dataset_ids}{dataset_ids}

\item{name}{name}

\item{tags}{tags}

\item{sort_key}{sort_key}

\item{ascending}{ascending}

\item{workspace}{workspace}
}
\description{

}
\details{

}
