% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{get_dataset}
\alias{get_dataset}
\title{get_dataset}
\usage{
get_dataset(name = NULL, workspace = NULL, id = NULL)
}
\arguments{
\item{name}{Name of the dataset. This parameter cannot be provided alongside id.}

\item{workspace}{Workspace under which the dataset with name name exists. If not provided, the current}

\item{id}{ID of the dataset. This parameter cannot be provided alongside name.}
}
\value{
:class:`~verta._dataset_versioning.dataset.Dataset`
}
\description{
Gets a dataset.
}
\details{
.. versionchanged:: 0.16.0 The dataset versioning interface was overhauled.
}
