% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{get_dataset_version}
\alias{get_dataset_version}
\title{get_dataset_version}
\usage{
get_dataset_version(id)
}
\arguments{
\item{id}{ID of the dataset version.}
}
\value{
:class:`verta._dataset_versioning.dataset_version.DatasetVersion`
}
\description{
Gets a dataset version.
}
\details{
.. versionchanged:: 0.16.0 The dataset versioning interface was overhauled.
}
