% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{get_endpoint}
\alias{get_endpoint}
\title{get_endpoint}
\usage{
get_endpoint(path = NULL, workspace = NULL, id = NULL)
}
\arguments{
\item{path}{Path of the Endpoint.}

\item{workspace}{Name of the workspace of the Endpoint.}

\item{id}{ID of the Endpoint. This parameter cannot be provided alongside path.}
}
\value{
:class:`~verta.endpoint._endpoint.Endpoint`
}
\description{
Retrieves an already created Endpoint. Only one of `path` or `id` can be provided.
}
\details{

}
