% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{get_experiment}
\alias{get_experiment}
\title{get_experiment}
\usage{
get_experiment(name = NULL, id = NULL)
}
\arguments{
\item{name}{Name of the Experiment.}

\item{id}{ID of the Experiment. This parameter cannot be provided alongside name.}
}
\value{
:class:`~verta._tracking.experiment.Experiment`
}
\description{
Retrieves an already created Experiment. Only one of `name` or `id` can be provided.
}
\details{

}
