% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{get_experiment_run}
\alias{get_experiment_run}
\title{get_experiment_run}
\usage{
get_experiment_run(name = NULL, id = NULL)
}
\arguments{
\item{name}{Name of the Experiment Run.}

\item{id}{ID of the Experiment Run. This parameter cannot be provided alongside name.}
}
\value{
:class:`verta._tracking.experimentrun.ExperimentRun`
}
\description{
Retrieves an already created Experiment Run. Only one of `name` or `id` can be provided.
}
\details{

}
