% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{get_or_create_registered_model}
\alias{get_or_create_registered_model}
\title{get_or_create_registered_model}
\usage{
get_or_create_registered_model(
  name = NULL,
  desc = NULL,
  labels = NULL,
  workspace = NULL,
  public_within_org = NULL,
  visibility = NULL,
  id = NULL
)
}
\arguments{
\item{name}{name}

\item{desc}{desc}

\item{labels}{labels}

\item{workspace}{workspace}

\item{public_within_org}{public_within_org}

\item{visibility}{visibility}

\item{id}{id}
}
\value{
:class:`~verta.registry._entities.model.RegisteredModel`
}
\description{
Attaches a registered_model to this Client.
}
\details{
If an accessible registered_model with name `name` does not already exist, it will be created
and initialized with specified metadata parameters. If such a registered_model does already exist,
it will be retrieved; specifying metadata parameters in this case will raise a warning.
}
