% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{get_or_create_repository}
\alias{get_or_create_repository}
\title{get_or_create_repository}
\usage{
get_or_create_repository(
  name = NULL,
  workspace = NULL,
  id = NULL,
  public_within_org = NULL,
  visibility = NULL
)
}
\arguments{
\item{name}{Name of the Repository. This parameter cannot be provided alongside id.}

\item{workspace}{Workspace under which the Repository with name name exists. If not provided, the}

\item{id}{ID of the Repository, to be provided instead of name.}

\item{public_within_org}{If creating a Repository in an organization's workspace: True}

\item{visibility}{Visibility to set when creating this repository. If not provided,}
}
\value{
:class:`verta._repository.Repository` Specified Repository.
}
\description{
Gets or creates a Repository by `name` and `workspace`, or gets a Repository by `id`.
}
\details{

}
