% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{get_project}
\alias{get_project}
\title{get_project}
\usage{
get_project(name = NULL, workspace = NULL, id = NULL)
}
\arguments{
\item{name}{Name of the Project.}

\item{workspace}{Workspace under which the Project with name name exists. If not provided, the current}

\item{id}{ID of the Project. This parameter cannot be provided alongside name.}
}
\value{
:class:`verta._tracking.project.Project`
}
\description{
Retrieves an already created Project. Only one of `name` or `id` can be provided.
}
\details{

}
