% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{get_registered_model_version}
\alias{get_registered_model_version}
\title{get_registered_model_version}
\usage{
get_registered_model_version(id)
}
\arguments{
\item{id}{ID of the Model Version.}
}
\value{
:class:`verta.registry._entities.modelversion.RegisteredModelVersion`
}
\description{
Retrieve an already created Model Version.
}
\details{

}
