% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{init_verta}
\alias{init_verta}
\title{Initialize Verta Client object}
\usage{
init_verta(
  HOST,
  method = "conda",
  conda = "auto",
  envname = "verta_reticulate",
  extra_packages = NULL,
  conda_python_version = "3.9.5",
  python_path = NULL
)
}
\arguments{
\item{HOST}{The}

\item{method}{- auto, virtualenv,conda or python -default conda}

\item{conda}{- path to conda executable}

\item{envname}{name of the conda environment}

\item{python_path}{path, optional}

\item{python}{one of [python,conda, miniconda, venv]}
}
\value{

}
\description{
by default using conda with the default settings is recommended.
This will create a verta_reticulate conda env, and the verta python client
will be installed there.
If the method is auto or python, we'll try to install verta using
pip or pip3. If python_path is not provided, we'll try to find it.
}
