% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{install_verta}
\alias{install_verta}
\title{Install Verta Library}
\usage{
install_verta(
  method = c("conda", "auto", "virtualenv", "python"),
  conda = "auto",
  envname = "verta_reticulate",
  extra_packages = NULL,
  conda_python_version = "3.9.5",
  python_path = NULL,
  ...
)
}
\arguments{
\item{method}{- auto, virtualenv, or conda}

\item{conda}{The path to a conda executable. Use "auto" to allow reticulate
to automatically find an appropriate conda binary. See
Finding Conda in reticulate  for more details.}

\item{envname}{The name, or full path, of the environment in which Python
packages are to be installed. When NULL (the default), the active environment
as set by the RETICULATE_PYTHON_ENV variable will be used; if that is unset,
then the r-reticulate environment will be used.}

\item{extra_packages}{Extra packages to install.}

\item{conda_python_version}{passed to reticulate::py_install}

\item{...}{passed down to reticulate::py_install}
}
\value{
Nothing
}
\description{
Install Verta Library
}
