% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{log_model}
\alias{log_model}
\title{log_model}
\usage{
log_model(run, model, overwrite = FALSE)
}
\arguments{
\item{model}{r object list with the following fields:}

\item{custom_modules}{custom modules.}
}
\description{
Logs a model to this Model Version.

workspace : str Verta workspace.
}
\section{workspace }{
 str:
Verta workspace.
}

