% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{log_run_hyperparameters}
\alias{log_run_hyperparameters}
\title{log_hyperparameters}
\usage{
log_run_hyperparameters(run, hyperparams, overwrite = FALSE)
}
\arguments{
\item{run}{The experiment run in question}

\item{hyperparams}{hyperparams}

\item{overwrite}{overwrite}
}
\description{
Logs potentially multiple hyperparameters to this Experiment Run.
}
\details{
Parameters
----------
hyperparameters : dict of str to {NULL, bool, float, int, str} Hyperparameters.
overwrite : bool, default FALSE Whether to allow overwriting an existing hyperparameter with key `key`.
}
\section{hyperparameters }{
 dict of str to {None, bool, float, int, str}:
Hyperparameters. overwrite : bool, default FALSE Whether to allow overwriting an existing hyperparameter with key `key`.
}

\section{overwrite }{
 bool, default False:
Whether to allow overwriting an existing hyperparameter with key `key`.
}

