% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_io.R
\name{predict_new_model}
\alias{predict_new_model}
\title{Predict Data From Saved Model}
\usage{
predict_new_model(file, data)
}
\arguments{
\item{file}{path to the file, holding the data}

\item{data}{newdata to predict on}
}
\value{

}
\description{
Predict Data From Saved Model
}
