% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{run_log_metric}
\alias{run_log_metric}
\title{run_log_metric}
\usage{
run_log_metric(run, key, value, overwrite = FALSE)
}
\arguments{
\item{key}{key}

\item{value}{value}

\item{overwrite}{overwrite}
}
\description{
Logs a metric to this Experiment Run.
}
\details{
If the metadatum of interest might recur, :meth:`.log_observation` should be used instead. Parameters
----------
key : str Name of the metric.
value : one of {NULL, bool, float, int, str} Value of the metric.
overwrite : bool, default FALSE Whether to allow overwriting an existing metric with key `key`.
}
\section{key }{
 str:
Name of the metric. value : one of {NULL, bool, float, int, str} Value of the metric. overwrite : bool, default FALSE Whether to allow overwriting an existing metric with key `key`.
}

\section{value }{
 one of {None, bool, float, int, str}:
Value of the metric. overwrite : bool, default FALSE Whether to allow overwriting an existing metric with key `key`.
}

\section{overwrite }{
 bool, default False:
Whether to allow overwriting an existing metric with key `key`.
}

