% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{set_experiment}
\alias{set_experiment}
\title{set_experiment}
\usage{
set_experiment(name = NULL, desc = NULL, tags = NULL, attrs = NULL, id = NULL)
}
\arguments{
\item{name}{Name of the Experiment. If no name is provided, one will be generated.}

\item{desc}{Description of the Experiment.}

\item{tags}{Tags of the Experiment.}

\item{attrs}{Attributes of the Experiment.}

\item{id}{ID of the Experiment. This parameter cannot be provided alongside name, and other
parameters will be ignored.}
}
\value{
:class:`verta._tracking.experiment.Experiment`
}
\description{
Attaches an Experiment under the currently active Project to this Client.
}
\details{
If an accessible Experiment with name `name` does not already exist under the currently
active Project, it will be created and initialized with specified metadata parameters. If
such an Experiment does already exist, it will be retrieved; specifying metadata parameters
in this case will raise a warning.
}
