% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{set_experiment_run}
\alias{set_experiment_run}
\title{set_experiment_run}
\usage{
set_experiment_run(
  name = NULL,
  desc = NULL,
  tags = NULL,
  attrs = NULL,
  id = NULL,
  date_created = NULL
)
}
\arguments{
\item{name}{name}

\item{desc}{desc}

\item{tags}{tags}

\item{attrs}{attrs}

\item{id}{id}

\item{date_created}{date_created}
}
\description{
Attaches an Experiment Run under the currently active Experiment to this Client.
}
\details{
If an accessible Experiment Run with name `name` does not already exist under the
currently active Experiment, it will be created and initialized with specified metadata
parameters. If such a Experiment Run does already exist, it will be retrieved; specifying
metadata parameters in this case will raise a warning. Parameters
----------
name : str, optional Name of the Experiment Run. If no name is provided, one will be generated.
desc : str, optional Description of the Experiment Run.
tags : list of str, optional Tags of the Experiment Run.
attrs : dict of str to {NULL, bool, float, int, str}, optional Attributes of the Experiment Run.
id : str, optional ID of the Experiment Run. This parameter cannot be provided alongside `name`, and other parameters will be ignored. Returns
-------
:class:`verta.tracking.entities.ExperimentRun` Raises
------
ValueError If an Experiment Run with `name` already exists, but metadata parameters are passed in.
AttributeError If an Experiment is not yet in progress.
}
\section{name }{
 str, optional:
Name of the Experiment Run. If no name is provided, one will be generated. desc : str, optional Description of the Experiment Run. tags : list of str, optional Tags of the Experiment Run. attrs : dict of str to {NULL, bool, float, int, str}, optional Attributes of the Experiment Run. id : str, optional ID of the Experiment Run. This parameter cannot be provided alongside `name`, and other parameters will be ignored.
}

\section{desc }{
 str, optional:
Description of the Experiment Run. tags : list of str, optional Tags of the Experiment Run. attrs : dict of str to {NULL, bool, float, int, str}, optional Attributes of the Experiment Run. id : str, optional ID of the Experiment Run. This parameter cannot be provided alongside `name`, and other parameters will be ignored.
}

\section{tags }{
 list of str, optional:
Tags of the Experiment Run. attrs : dict of str to {NULL, bool, float, int, str}, optional Attributes of the Experiment Run. id : str, optional ID of the Experiment Run. This parameter cannot be provided alongside `name`, and other parameters will be ignored.
}

\section{attrs }{
 dict of str to {None, bool, float, int, str}, optional:
Attributes of the Experiment Run. id : str, optional ID of the Experiment Run. This parameter cannot be provided alongside `name`, and other parameters will be ignored.
}

\section{id }{
 str, optional:
ID of the Experiment Run. This parameter cannot be provided alongside `name`, and other parameters will be ignored.
}

