% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{set_project}
\alias{set_project}
\title{set_project}
\usage{
set_project(
  name = NULL,
  desc = NULL,
  tags = NULL,
  attrs = NULL,
  workspace = NULL,
  public_within_org = NULL,
  visibility = NULL,
  id = NULL
)
}
\arguments{
\item{name}{Name of the Project. If no name is provided, one will be generated.}

\item{desc}{Description of the Project.}

\item{tags}{Tags of the Project.}

\item{attrs}{Attributes of the Project.}

\item{workspace}{Workspace under which the Project with name name exists. If not provided, the current
user's personal workspace will be used.}

\item{public_within_org}{If creating a Project in an organization's workspace: True for}

\item{visibility}{Visibility to set when creating this project. If not provided, an}

\item{id}{ID of the Project. This parameter cannot be provided alongside name, and other}
}
\value{
:class:`~verta._tracking.project.Project`
}
\description{
Attaches a Project to this Client.
}
\details{
If an accessible Project with name `name` does not already exist, it will be created
and initialized with specified metadata parameters. If such a Project does already exist,
it will be retrieved; specifying metadata parameters in this case will raise a warning. If an Experiment is already attached to this Client, it will be detached.
}
