% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_api_scaffold.R
\name{set_workspace}
\alias{set_workspace}
\title{set_workspace}
\usage{
set_workspace(workspace)
}
\arguments{
\item{workspace}{Verta workspace.}
}
\description{
Sets the active workspace for this client instance.
}
\details{
.. versionadded:: 0.17.0 Parameters
----------
workspace : str Verta workspace.
}
\section{workspace }{
 str:
Verta workspace.
}

