% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_io.R
\name{createRandString}
\alias{createRandString}
\title{Title}
\usage{
createRandString(
  letter_len1 = 5,
  digit_len = 4,
  letter_len2 = 1,
  add_docker_substring = T,
  add_rdata = F
)
}
\arguments{
\item{add_rdata}{}
}
\value{

}
\description{
Title
}
